pyproject.toml
[tool.poetry]
name = "play-write"
version = "0.1.0"
description = ""
authors = ["mikhael <sur.mikhael@gmail.com>"]
readme = "README.md"
packages = [{include = "play_write"}]

[tool.poetry.dependencies]
python = "^3.12"
pytest-playwright = "^0.4.4"
python-dotenv = "^1.0.1"
black = "^24.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
